+++ text_to_speech/setup.py
@@ -15,4 +15,3 @@ setup(
         "openai==1.78.1 ",
     ],
 )
-
