+++ text_to_speech/src/robot_vocalizer/robot_vocalizer.py
@@ -1,6 +1,6 @@
 import logging
-import tempfile
 import subprocess
+import tempfile
 
 from deepgram import (
     DeepgramClient,
@@ -18,9 +18,9 @@ class RobotVocalizer:
 
     def vocalize(self, text: str):
         with tempfile.NamedTemporaryFile(delete=False) as tmp_file:
-            response = self.deepgram.speak.rest.v("1").save(tmp_file.name, {"text": text}, self.options)
+            response = self.deepgram.speak.rest.v("1").save(
+                tmp_file.name, {"text": text}, self.options
+            )
             logger.debug(f"Vocalized response: {response.to_json(indent=4)}")
             # subprocess.run(["afplay", tmp_file.name]) # Mac
             subprocess.run(["ffplay", tmp_file.name])
-
-
